+++ data/categories.ts
@@ -1,4 +1,4 @@
-import type { CategoriesGroup, Category } from "./types"
+import type { CategoriesGroup, Category } from "../@types/types"
 // import images from './images';
 
 export const CATEGORIES: CategoriesGroup[] = [
