+++ pages/index.page.tsx
@@ -1,9 +1,9 @@
 import { Container, Heading, VStack, Wrap } from "@yamada-ui/react"
 import type { InferGetStaticPropsType, NextPage } from "next"
+import type { CategoriesGroup } from "@types/types"
 import { CategoryCard } from "components/layouts/category-card"
 import { useI18n } from "contexts/i18n-context"
 import { CATEGORIES } from "data/categories"
-import type { CategoriesGroup } from "data/types"
 import { AppLayout } from "layouts/app-layout"
 
 type PageProps = InferGetStaticPropsType<typeof getStaticProps>
